src/ai_core/project_builder.py
"""Module for project_builder."""

from pathlib import Path
from typing import Dict

import toml


        class ProjectBuilder:
        """Handles the creation of project files and directories."""
        """
        """
        This class is responsible for creating individual project components
        and ensuring they follow the required structure and format.
        """
        """
        """
        """
            def __init__(self, config: ProjectConfig, root_dir: Path):
            """Initialize the project builder."""
            """
            """
            Args:
            config: Project configuration
            root_dir: Project root directory
            """
            """
            self.config = config"""
            """
            self.root_dir = root_dir

                def create_file(self, path: Path, content: str) -> None:
                """Create a file with the given content."""
                """
                """
                Args:
                path: Path to the file
                content: File content
                """
                """
                path.parent.mkdir(parents=True, exist_ok=True)"""
                """
                path.write_text(content)

                    def create_pyproject_toml(self) -> None:
                    """Create Poetry configuration file."""
                    poetry_config = {
                    "tool": {
                    "poetry": {
                    "name": self.config.name,
                    "version": self.config.version,
                    "description": self.config.description,
                    "authors": ["AI Assistant <ai@example.com>"],
                    "readme": "README.md",
                    "dependencies": self._get_dependencies(),
                    "dev-dependencies": self._get_dev_dependencies(),
                    "scripts": self._get_script_definitions(),
                    }
                    },
                    "build-system": {
                    "requires": ["poetry-core>=1.0.0"],
                    "build-backend": "poetry.core.masonry.api",
                    },
                    }

                    self.create_file(self.root_dir / "pyproject.toml", toml.dumps(poetry_config))

                        def _get_dependencies(self) -> Dict[str, str]:
                        """Generate dependency dictionary with versions."""
                        deps = {"python": self.config.python_version}
                            for dep in self.config.base_dependencies:
                            deps[dep] = "*"
                        return deps

                            def _get_dev_dependencies(self) -> Dict[str, str]:
                            """Generate development dependency dictionary."""
                        return {dep: "*" for dep in self.config.dev_dependencies}

                            def _get_script_definitions(self) -> Dict[str, str]:
                            """Define Poetry scripts for common operations."""
                        return {
                        "start": f"python -m {self.config.name}",
                        "test": "pytest",
                        "test-cov": "pytest --cov",
                        "lint": "pylint src tests",
                        "format": "black src tests",
                        "format-check": "black --check src tests",
                        "typecheck": "mypy src",
                        "sort-imports": "isort .",
                        "check": "python -m scripts.check_all",
                        }
